moved away from in-memory factory service for Users for now
