+++ src/domain/user/user.module.ts
@@ -6,19 +6,17 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { UserResolverFields } from './user.resolver.fields';
 import { IsUserAlreadyExistConstraint } from '@utils/validation/constraints/user.exists.constraint';
-import { UserFactoryService } from './user.factory.service';
 import { UserResolverMutations } from './user.resolver.mutations';
 
 @Module({
   imports: [ProfileModule, TypeOrmModule.forFeature([User])],
   providers: [
     UserService,
-    UserFactoryService,
     UserResolverMutations,
     UserResolverQueries,
     UserResolverFields,
     IsUserAlreadyExistConstraint,
   ],
-  exports: [UserService, UserFactoryService],
+  exports: [UserService],
 })
 export class UserModule {}
